removed react as it was created stupid issues that made me waste a lot of time
