+++ assets/style.css
@@ -1,0 +1,285 @@
+body {
+    margin: 0;
+    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+      'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+      sans-serif;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+    background-color: rgb(36, 36, 36);
+    color: #4dfa87;
+  }
+  
+  code {
+    font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+      monospace;
+  }
+  
+
+.App {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    flex-wrap: wrap;
+  }
+  
+  #download-button {
+    align-self: flex-end;
+    width: 100px;
+    font-size: 15px;
+    height: 20px;
+    top: 0;
+    right: 0;
+    margin-right: 20px;
+    display: block;
+  }
+  
+  .sections-and-editor {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+  }
+  
+  #markdown-preview {
+    background-color: #3b3b3b;
+    margin-left: 20px;
+    width: 400px;
+    height: 300px;
+  }
+  
+  .markdown-preview-h2 {
+    margin-left: 80px;
+  }
+  
+  textarea {
+    background-color: #3b3b3b;
+    color: #abffc7;
+    width: 300px;
+    height: 300px;
+    margin-left: 20px;
+    resize: none;
+  }
+  textarea:focus {
+    outline: none !important;
+    border: 1px solid #abffc7
+  }
+  
+  .editor-div {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+  }
+  
+  h2 {
+    margin-bottom: 0;
+    margin-top: 0;
+  }
+  
+  .readme-sections-h2 {
+    margin-left: 20px;
+  }
+  
+  h1 {
+    margin-left: 50px;
+    margin-right: 50px;
+  }
+  
+  input {
+    background-color: gray;
+    color: #4dfa87;
+    width: 300px;
+    height: 30px;
+    font-size: 20px;
+    margin-bottom: 5px;
+  }
+  input:focus {
+    outline: none !important;
+    border: 1px solid grey
+  }
+  
+  ::placeholder {
+    color: rgb(54, 54, 54);
+    opacity: 1; /* Firefox */
+  }
+  
+  ::-ms-input-placeholder { /* Edge 12 -18 */
+    color: rgb(54, 54, 54);
+  }
+  
+  textarea::placeholder {
+    color: #4dfa87;
+    opacity: 1; /* Firefox */
+  }
+  
+  textarea::-ms-input-placeholder { /* Edge 12 -18 */
+    color: rgb(54, 54, 54);
+  }
+  
+  form {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+  }
+  
+  span{
+    color: gray;
+    margin-top: -20px;
+    margin-bottom: 30px;
+    margin-left: 50px;
+    margin-right: 50px;
+  }
+  
+  #readme-form{
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+  }
+  
+  button {
+    margin-top: 20px;
+    background-color: #4dfa87;
+    width: 200px;
+    height: 55px;
+    font-size: 20px;
+    border-radius: 10px;
+    display: none;
+    border: 0;
+    align-items: center;
+  }
+  button:hover {
+    background-color: #abffc7;
+    cursor: pointer;
+  }
+  
+  a{
+    text-decoration: none;
+  }
+  
+  .contact-button {
+    display: flex;
+    margin-left: 10px;
+  }
+  
+  #custom-input-button {
+    display: flex;
+  }
+  
+  .custom-contact-buttons {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: center;
+    margin-bottom: 20px;
+  }
+  
+  .start-span {
+    width: 500px;
+  }
+  
+  .main-content-div {
+    display: flex;
+    flex-direction: column;
+  }
+  
+  #section-editor {
+    height: 300px;
+  }
+  
+  #new-section-input {
+    width: 200px;
+  }
+  
+  #new-section-button {
+    display: inline;
+    width: 30px;
+    height: 32px;
+    border-radius: 0;
+  }
+  
+  #new-section-form {
+    justify-content: center;
+    align-items: center;
+    display: flex;
+    flex-direction: row;
+    align-self: flex-start;
+  }
+  
+  .section-span {
+    display: inline-block;
+    font-size: 20px;
+    color: #4dfa87;
+    padding: 10px;
+    padding-top: 2px;
+    padding-bottom: 2px;
+    background-color: grey;
+    margin: 0;
+    width: 140px;
+  }
+  .section-span:hover {
+    background-color: gray;
+    cursor: pointer;
+  }
+  
+  .section-button {
+    display: inline;
+    width: 70px;
+    height: 29px;
+    border-radius: 0;
+  }
+  
+  #sections-div {
+    background-color: #abffc7;
+    display: flex;
+    flex-direction: column;
+    overflow-y: scroll;
+    background-color: #3b3b3b;
+    padding: 5px;
+    width: 250px;
+    align-items: center;
+    height: 300px;
+  }
+  #new-input-form input {
+    background-color: #abffc7;
+    width: 180px;
+    color: black;
+  }
+  
+  /* SCROLLBAR STYLES */
+  /* width */
+  ::-webkit-scrollbar {
+    width: 10px;
+  }
+  
+  /* Track */
+  ::-webkit-scrollbar-track {
+    background: #3b3b3b;
+  }
+  
+  /* Handle */
+  ::-webkit-scrollbar-thumb {
+    background: #4dfa87;
+  }
+  
+  /* Handle on hover */
+  ::-webkit-scrollbar-thumb:hover {
+    background: #abffc7;
+  }
+  
+  #editor {
+    width: 500px;
+    height: 500px;
+  }
+  
+  /* highlight colors */
+  ::-moz-selection { /* Code for Firefox */
+    color: #3b3b3b;
+    background: #abffc7;
+  }
+  
+  ::selection {
+    color: #3b3b3b;
+    background: #abffc7;
+  }
